split CLI from main package lib and use md5 hash to determine if tool
